Fix AdapterFusionConfig serialization. Refactor adapter tests.
