+++ src/transformers/adapter_config.py
@@ -6,7 +6,7 @@ from dataclasses import FrozenInstanceError, asdict, dataclass, field, is_datacl
 from os.path import isfile
 from typing import List, Optional, Union
 
-from .adapter_utils import AdapterType, DataclassJSONEncoder, get_adapter_config_hash, resolve_adapter_config
+from .adapter_utils import AdapterType, get_adapter_config_hash, resolve_adapter_config
 
 
 logger = logging.getLogger(__name__)
@@ -249,8 +249,6 @@ class ModelAdaptersConfig:
     def to_dict(self):
         output_dict = {}
         output_dict["adapters"] = copy.deepcopy(self.adapters)
-        # make sure all config objects are serializable
-        output_dict["config_map"] = json.loads(json.dumps(self.config_map, cls=DataclassJSONEncoder))
         return output_dict
