+++ tests/test_adapter_fusion_loading.py
@@ -95,3 +95,18 @@ class AdapterFusionModelTest(unittest.TestCase):
                         output2 = model2(in_data, adapter_names=[[name1, name2]])
                         self.assertEqual(len(output1), len(output2))
                         self.assertTrue(torch.equal(output1[0], output2[0]))
+
+    def test_model_config_serialization(self):
+        """PretrainedConfigurations should not raise an Exception when serializing the config dict
+
+        See, e.g., PretrainedConfig.to_json_string()
+        """
+        for model_class in self.model_classes:
+            for k, v in ADAPTERFUSION_CONFIG_MAP.items():
+                model_config = model_class.config_class
+                model = model_class(model_config())
+                model.add_adapter("test1", AdapterType.text_task)
+                model.add_adapter("test2", AdapterType.text_task)
+                model.add_fusion(["test1", "test2"], adapter_fusion_config=v)
+                # should not raise an exception
+                model.config.to_json_string()
